feat: created frontend component to filter by genre
